[webhook] Move as own root package
